0210/study29/app1/pyproject.toml
[project]
name = "app1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.6",
    "kafka-python>=2.3.0",
]
